Move `JpegStream` and `JpxStream` to their own files
